requirements.txt
Flask>=2.0.1
python-dotenv>=0.19.0
Telethon>=1.24.0
